+++ tests/e2e/CHANGELOG.md
@@ -1,5 +1,33 @@
 # e2e
 
+## 0.0.1-beta.2
+
+### Patch Changes
+
+- [`b528b32`](https://github.com/kubb-labs/plugins/commit/b528b3226d796a6aab5f1f6d45b575921da1341b) Thanks [@stijnvanhulle](https://github.com/stijnvanhulle)! - sync between core with same version
+
+- Updated dependencies [[`b528b32`](https://github.com/kubb-labs/plugins/commit/b528b3226d796a6aab5f1f6d45b575921da1341b)]:
+  - @kubb/plugin-client@5.0.0-beta.22
+  - @kubb/plugin-cypress@5.0.0-beta.22
+  - @kubb/plugin-faker@5.0.0-beta.22
+  - @kubb/plugin-msw@5.0.0-beta.22
+  - @kubb/plugin-react-query@5.0.0-beta.22
+  - @kubb/plugin-ts@5.0.0-beta.22
+  - @kubb/plugin-zod@5.0.0-beta.22
+
+## 0.0.1-beta.1
+
+### Patch Changes
+
+- Updated dependencies [[`234a4d7`](https://github.com/kubb-labs/plugins/commit/234a4d7c9dccb1f756447e8d70d4a5bec4dcf72f)]:
+  - @kubb/plugin-client@5.0.0-beta.15
+  - @kubb/plugin-cypress@5.0.0-beta.15
+  - @kubb/plugin-faker@5.0.0-beta.15
+  - @kubb/plugin-msw@5.0.0-beta.15
+  - @kubb/plugin-react-query@5.0.0-beta.15
+  - @kubb/plugin-ts@5.0.0-beta.15
+  - @kubb/plugin-zod@5.0.0-beta.15
+
 ## 0.0.1-beta.0
 
 ### Patch Changes